requirements.txt
numpy
matplotlib
Pillow
serial